+++ CHANGELOG.md
@@ -11,4 +11,4 @@ This change log will document the notable changes to this project in this file a
 ### Fixed
 
 ### Changed
-- Changed name of package to NIPTool
+- Changed name of package to statina